+++ storage/downloader-file-system.go
@@ -10,7 +10,7 @@ type FileSystemDownloader struct {
 	Fs FileSystem
 }
 
-func (d FileSystemDownloader) Download(gameId int, add manifest.FileAction) (io.ReadCloser, int64, string, error) {
+func (d FileSystemDownloader) Download(gameId int64, add manifest.FileAction) (io.ReadCloser, int64, string, error) {
 	handle, size, err := d.Fs.DownloadFile(gameId, add.Kind, add.Name)
 	return handle, size, add.Name, err
 }
