+++ storage/storage-interface.go
@@ -22,9 +22,9 @@ type Storage interface {
 	LoadSource() (*Source, error)
 	RemoveActions() error
 	RemoveSource() error
-	AddGame(gameId int) error
-	RemoveGame(gameId int) error
-	UploadFile(source io.ReadCloser, gameId int, kind string, name string, expectedSize int64) (string, error)
-	RemoveFile(gameId int, kind string, name string) error
-	DownloadFile(gameId int, kind string, name string) (io.ReadCloser, int64, error)
+	AddGame(gameId int64) error
+	RemoveGame(gameId int64) error
+	UploadFile(source io.ReadCloser, gameId int64, kind string, name string, expectedSize int64) (string, error)
+	RemoveFile(gameId int64, kind string, name string) error
+	DownloadFile(gameId int64, kind string, name string) (io.ReadCloser, int64, error)
 }
